+++ src/pages/sunshine/[entity]/[id]/costs-and-tariffs.tsx
@@ -1,4 +1,4 @@
-import { Trans, t } from "@lingui/macro";
+import { t, Trans } from "@lingui/macro";
 import { GetServerSideProps } from "next";
 import ErrorPage from "next/error";
 import { useRouter } from "next/router";
@@ -7,8 +7,8 @@ import React, { useState } from "react";
 import CardGrid from "src/components/card-grid";
 import { DetailPageBanner } from "src/components/detail-page/banner";
 import {
-  DetailsPageLayout,
   DetailsPageHeader,
+  DetailsPageLayout,
   DetailsPageSubtitle,
   DetailsPageTitle,
 } from "src/components/detail-page/layout";
@@ -171,6 +171,9 @@ const NetworkCosts = (props: Extract<Props, { status: "found" }>) => {
           sx={{ gridArea: "trend" }}
           peerGroup={peerGroup}
           updateDate={updateDate}
+          operatorId={props.id}
+          operatorLabel={operatorLabel}
+          networkCosts={props.costsAndTariffs.networkCosts}
         />
       </CardGrid>
     </>
@@ -261,6 +264,9 @@ const EnergyTariffs = (props: Extract<Props, { status: "found" }>) => {
           sx={{ gridArea: "trend" }}
           peerGroup={peerGroup}
           updateDate={updateDate}
+          operatorId={props.id}
+          operatorLabel={operatorLabel}
+          netTariffs={props.costsAndTariffs.netTariffs}
         />
       </CardGrid>
     </>
@@ -351,6 +357,9 @@ const NetTariffs = (props: Extract<Props, { status: "found" }>) => {
           sx={{ gridArea: "trend" }}
           peerGroup={peerGroup}
           updateDate={updateDate}
+          operatorId={props.id}
+          operatorLabel={operatorLabel}
+          netTariffs={props.costsAndTariffs.netTariffs}
         />
       </CardGrid>
     </>